flights/requirements.txt
Flask>=2.1.3
flasgger>=0.9.7.1
Flask-Cors>=5.0.0
opentelemetry-distro
opentelemetry-exporter-otlp
python-json-logger
# flask-pypprof
pyroscope-io